app/requirements.txt
fastapi==0.101.0
pydantic==2.1.1
uvicorn==0.23.2
requests==2.31.0
openai==0.27.8
python-dotenv==1.0.0
ipdb==0.13.13
jedi<0.18.0
